requirements.txt
pymongo[srv]
python-dotenv
certifi
browser_cookie3
requests
gunicorn
Flask
flask-cors